Avoid downloading the large zip file if we already have it.
There is no point in downloading the large file if it is already present
on the disc. It's possible that it might have failed to fully download,
but the unzip will at least tell you that. The unzip could also protect
against extracting the file again, but the unzip is cheap compared the
4GB download of a zip.