Move docker build to a subdirectory; upgrade wine to 4.0.2.
When the `docker build` was being run, it did so in the root directory,
which contains the downloaded zip file and the extracted contents of
that zip, in addition to the snapshots it required. This meant that the
build was transferring about 11GB of data to the docker daemon to make
available to the build system - not a fast or useful operation.

Moving the docker build to happen in a separate directory, which
contains only the things that are needed by the Docker build, means that
we only transfer about 2.8GB to the docker daemon, and it runs a lot
faster.

To make this work, the snapshots need to be created in a different
place so that they may be accessed by the docker, but aren't in the
same location as the builds. For ease, I've just symlink'd these
directories in the makefile.

Wine version 4.0 was previously required - but this does not appear
to install at the present time (August 2019). Upgrading the version
required to 4.0.2 in the Makefile allowed this to work.